refactor: water heating using exhaust is a processHeat calculator not steam
